Clean up and refactor.
- Removed some unused images in the public folder
- Extracted copy from index.html into yaml files in data/
- Build homepage from index.haml instead of statically
- Condense Rack logic to events.rb, homepage.rb, and config.ru

I tried to use map "/" { run HomePage.new } to route just root to the
homepage, but that makes everything starting with "/" get served by
the HomePage app. So I used a lambda in place to test the request path
and route to HomePage if and only if the PATH_INFO is exactly "/".
Everything else will be passed through to NotFound.